requirements.txt
jstyleson>=0.0.2
GitPython>=3.1.31
typer>=0.9.0